install_openvas.sh
This script is used to install and setup the Greenbone Vulnerability Manager (GVM), which is a vulnerability assessment tool. It does this by first adding the GVM repository to the system, then updating the package lists. Next, it installs GVM and creates a new admin user for the tool. After that, it starts the GVM services (scanner, manager, and feed updater) and updates the NVT and SCAP data, which are the vulnerabilities and security standards used by GVM to perform assessments. Overall, this script automates the process of installing and configuring GVM on a Linux system.